requirements.txt
# Core dependencies
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
ollama>=0.1.0

# Database
# Note: sqlite3 is built into Python

# Build tools (for Rust compilation)
maturin>=1.4.0

# System Tray & Notifications
pystray>=0.19.0
pillow>=10.0.0

# PDF Reports & Forensics
fpdf2>=2.7.0

# Process Tree Visualization
networkx>=3.0
streamlit-agraph>=0.0.45

# Audio Alerts
pyttsx3>=2.90

# GeoIP (for network visualization)
geoip2>=4.7.0
maxminddb>=2.3.0

# Testing (APEX)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0

# 3D Visualization
pydeck>=0.8.0
pydeck-carto>=0.1.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# YAML Rules
pyyaml>=6.0.0

# Performance Monitoring
psutil>=5.9.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
humanize>=4.8.0
